Refactor Boards to the TS
