--- src/resources/boards/board.model.js
@@ -1,18 +1,0 @@
-class Board {
-  constructor({
-    id,
-    title,
-    columns
-  } = {}) {
-    this.id = id;
-    this.title = title;
-    this.columns = columns;
-  }
-
-  static toResponse(board) {
-    const { id, title, columns } = board;
-    return { id, title, columns };
-  }
-}
-
-module.exports = Board;
